requirements.txt
streamlit==1.35.0
groq==0.9.0
langchain==0.2.6
langchain-groq==0.1.6
langchain-community==0.2.6
chromadb==0.5.3
sentence-transformers==3.0.1
tavily-python==0.3.3
fpdf2==2.7.9
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0